+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased](https://github.com/orisai/nette-monolog/compare/1.2.0...HEAD)
 
+### Added
+
+- Tracy panel design
+
 ## [1.2.0](https://github.com/orisai/nette-monolog/compare/1.1.0...1.2.0)
 
 ### Added